Add files via upload
In this Project, I was tasked with analyzing any dataset of my choosing. I chose to find the 
best cities to live in from data collected in 2022. This was a more visual and deeper dive where I 
used several visualizations of different variables to find the best results and why some cities are 
classified as better to live in than others.