Add files via upload
IBM gathered this dataset, which is the data about employee attrition and performance. I 
decided to focus on the data related to finance, such as monthly income and anything that could 
impact it, such as age, hourly rate, years with the company, years since the last promotion, and 
years in the current position. Another factor that could be considered is “Years with current 
manager”; having a healthy work environment can lead to a lower attrition rate and higher 
performance, although measuring such numbers cannot be easily proven to have a direct impact 
on an employee’s performance, it can be something interesting to consider. However, if one considers environmental satisfaction, job satisfaction, performance rating, and years in a current 
role and compares it to “Years with current manager,” it may lead to an indirect relationship.