Add files via upload
I was requested to make a simple analysis of any data of my choosing; in this case, I chose a dataset containing income and demographics data. I also sought to answer if maybe experience could be a reason why more mature people may occupy higher-paying jobs, as well as if gender was relative as well.