Add files via upload
A stadium requested their data to be analyzed with the goal of getting recommendations to increase their sales. For that, I had to analyze what their peak days were and why to make a concise recommendation. I was happy with my work, but there could be a thing or so that could be further analyzed.